requirements.txt
docopt==0.6.2
PyYAML==6.0
pytest==7.1.2
